Converted Gx use from fixed point to floating point math

+++ Stars/source/Level.cpp
@@ -130,7 +130,7 @@ void Level::EndDrawPathHandler(const LevelInteractor& sender, const LevelInterac
 	if (Star* star = m_xGame.GetStar()) {
 		if (path.count > 0) {
 			IwAssertMsg(MYAPP, star->IsDragging(), ("Star is not being dragged. Something's wrong!"));
-			star->FollowPath(path.count, path.samplepos, Configuration::GetInstance().PathSpeed);
+			star->FollowPath(path.count, path.samplepos, (float)Configuration::GetInstance().PathSpeed);
 		}
 	}
 }
